implemented provider for store in order to create store per each request to server, implemented saving store to local-storage
