+++ src/app/page.tsx
@@ -1,113 +1,641 @@
+'use client'
+
 import Image from 'next/image'
+import Link from 'next/link'
+import {
+  ChevronLeft,
+  ChevronRight,
+  Copy,
+  CreditCard,
+  File,
+  LineChart,
+  ListFilter,
+  MoreVertical,
+  Package,
+  Package2,
+  PanelLeft,
+  Search,
+  Settings,
+  ShoppingCart,
+  Truck,
+  Users2,
+} from 'lucide-react'
+// import { Badge } from '@/components/ui/badge'
+// import {
+//   Breadcrumb,
+//   BreadcrumbItem,
+//   BreadcrumbLink,
+//   BreadcrumbList,
+//   BreadcrumbPage,
+//   BreadcrumbSeparator,
+// } from '@/components/ui/breadcrumb'
+import { Button } from '@/components/ui/button'
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+} from '@/components/ui/card'
+import {
+  DropdownMenu,
+  DropdownMenuCheckboxItem,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from '@/components/ui/dropdown-menu'
+import { Input } from '@/components/ui/input'
+// import {
+//   Pagination,
+//   PaginationContent,
+//   PaginationItem,
+// } from '@/components/ui/pagination'
+// import { Progress } from '@/components/ui/progress'
+// import { Separator } from '@/components/ui/separator'
+// import { Sheet, SheetContent, SheetTrigger } from '@/components/ui/sheet'
+// import {
+//   Table,
+//   TableBody,
+//   TableCell,
+//   TableHead,
+//   TableHeader,
+//   TableRow,
+// } from '@/components/ui/table'
+// import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+} from '@/components/ui/tooltip'
+import { Sidebar } from '@/components/organisms/sidebar'
+import { Header } from '@/components/organisms/header'
+import { useEffect, useState } from 'react'
+import { useUserStore } from '@/store/user-store-provider'
+import { Code } from '@/utils/inter-layer-object'
 
 export default function Home() {
+  const { me, setSignedIn } = useUserStore((state) => state)
+
+  useEffect(() => {
+    const meRequest = async () => {
+      await me()
+    }
+    meRequest()
+  }, [me, setSignedIn])
+
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      <div className="z-10 w-full max-w-5xl items-center justify-between font-mono text-sm lg:flex">
-        <p className="fixed left-0 top-0 flex w-full justify-center border-b border-gray-300 bg-gradient-to-b from-zinc-200 pb-6 pt-8 backdrop-blur-2xl dark:border-neutral-800 dark:bg-zinc-800/30 dark:from-inherit lg:static lg:w-auto  lg:rounded-xl lg:border lg:bg-gray-200 lg:p-4 lg:dark:bg-zinc-800/30">
-          Get started by editing&nbsp;
-          <code className="font-mono font-bold">src/app/page.tsx</code>
-        </p>
-        <div className="fixed bottom-0 left-0 flex h-48 w-full items-end justify-center bg-gradient-to-t from-white via-white dark:from-black dark:via-black lg:static lg:size-auto lg:bg-none">
-          <a
-            className="pointer-events-none flex place-items-center gap-2 p-8 lg:pointer-events-auto lg:p-0"
-            href="https://vercel.com?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            By{' '}
-            <Image
-              src="/vercel.svg"
-              alt="Vercel Logo"
-              className="dark:invert"
-              width={100}
-              height={24}
-              priority
-            />
-          </a>
-        </div>
-      </div>
+    <div className="flex min-h-screen w-full flex-col bg-muted/40">
+      <Header />
+      <Sidebar />
 
-      <div className="relative z-[-1] flex place-items-center before:absolute before:h-[300px] before:w-full before:-translate-x-1/2 before:rounded-full before:bg-gradient-radial before:from-white before:to-transparent before:blur-2xl before:content-[''] after:absolute after:-z-20 after:h-[180px] after:w-full after:translate-x-1/3 after:bg-gradient-conic after:from-sky-200 after:via-blue-200 after:blur-2xl after:content-[''] before:dark:bg-gradient-to-br before:dark:from-transparent before:dark:to-blue-700 before:dark:opacity-10 after:dark:from-sky-900 after:dark:via-[#0141ff] after:dark:opacity-40 sm:before:w-[480px] sm:after:w-[240px] before:lg:h-[360px]">
-        <Image
-          className="relative dark:drop-shadow-[0_0_0.3rem_#ffffff70] dark:invert"
-          src="/next.svg"
-          alt="Next.js Logo"
-          width={180}
-          height={37}
-          priority
-        />
-      </div>
-
-      <div className="mb-32 grid text-center lg:mb-0 lg:w-full lg:max-w-5xl lg:grid-cols-4 lg:text-left">
-        <a
-          href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <h2 className="mb-3 text-2xl font-semibold">
-            Docs{' '}
-            <span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
-              -&gt;
-            </span>
-          </h2>
-          <p className="m-0 max-w-[30ch] text-sm opacity-50">
-            Find in-depth information about Next.js features and API.
-          </p>
-        </a>
-
-        <a
-          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <h2 className="mb-3 text-2xl font-semibold">
-            Learn{' '}
-            <span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
-              -&gt;
-            </span>
-          </h2>
-          <p className="m-0 max-w-[30ch] text-sm opacity-50">
-            Learn about Next.js in an interactive course with&nbsp;quizzes!
-          </p>
-        </a>
-
-        <a
-          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <h2 className="mb-3 text-2xl font-semibold">
-            Templates{' '}
-            <span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
-              -&gt;
-            </span>
-          </h2>
-          <p className="m-0 max-w-[30ch] text-sm opacity-50">
-            Explore starter templates for Next.js.
-          </p>
-        </a>
-
-        <a
-          href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <h2 className="mb-3 text-2xl font-semibold">
-            Deploy{' '}
-            <span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
-              -&gt;
-            </span>
-          </h2>
-          <p className="m-0 max-w-[30ch] text-balance text-sm opacity-50">
-            Instantly deploy your Next.js site to a shareable URL with Vercel.
-          </p>
-        </a>
-      </div>
-    </main>
+      {/*<div className="flex flex-col sm:gap-4 sm:py-4 sm:pl-14">*/}
+      {/*  <header className="sticky top-0 z-30 flex h-14 items-center gap-4 border-b bg-background px-4 sm:static sm:h-auto sm:border-0 sm:bg-transparent sm:px-6">*/}
+      {/*    <Sheet>*/}
+      {/*      <SheetTrigger asChild>*/}
+      {/*        <Button size="icon" variant="outline" className="sm:hidden">*/}
+      {/*          <PanelLeft className="h-5 w-5" />*/}
+      {/*          <span className="sr-only">Toggle Menu</span>*/}
+      {/*        </Button>*/}
+      {/*      </SheetTrigger>*/}
+      {/*      <SheetContent side="left" className="sm:max-w-xs">*/}
+      {/*        <nav className="grid gap-6 text-lg font-medium">*/}
+      {/*          <Link*/}
+      {/*            href="#"*/}
+      {/*            className="group flex h-10 w-10 shrink-0 items-center justify-center gap-2 rounded-full bg-primary text-lg font-semibold text-primary-foreground md:text-base"*/}
+      {/*          >*/}
+      {/*            <Package2 className="h-5 w-5 transition-all group-hover:scale-110" />*/}
+      {/*            <span className="sr-only">Acme Inc</span>*/}
+      {/*          </Link>*/}
+      {/*          <Link*/}
+      {/*            href="#"*/}
+      {/*            className="flex items-center gap-4 px-2.5 text-muted-foreground hover:text-foreground"*/}
+      {/*          >*/}
+      {/*            <Home className="h-5 w-5" />*/}
+      {/*            Dashboard*/}
+      {/*          </Link>*/}
+      {/*          <Link*/}
+      {/*            href="#"*/}
+      {/*            className="flex items-center gap-4 px-2.5 text-foreground"*/}
+      {/*          >*/}
+      {/*            <ShoppingCart className="h-5 w-5" />*/}
+      {/*            Orders*/}
+      {/*          </Link>*/}
+      {/*          <Link*/}
+      {/*            href="#"*/}
+      {/*            className="flex items-center gap-4 px-2.5 text-muted-foreground hover:text-foreground"*/}
+      {/*          >*/}
+      {/*            <Package className="h-5 w-5" />*/}
+      {/*            Products*/}
+      {/*          </Link>*/}
+      {/*          <Link*/}
+      {/*            href="#"*/}
+      {/*            className="flex items-center gap-4 px-2.5 text-muted-foreground hover:text-foreground"*/}
+      {/*          >*/}
+      {/*            <Users2 className="h-5 w-5" />*/}
+      {/*            Customers*/}
+      {/*          </Link>*/}
+      {/*          <Link*/}
+      {/*            href="#"*/}
+      {/*            className="flex items-center gap-4 px-2.5 text-muted-foreground hover:text-foreground"*/}
+      {/*          >*/}
+      {/*            <LineChart className="h-5 w-5" />*/}
+      {/*            Settings*/}
+      {/*          </Link>*/}
+      {/*        </nav>*/}
+      {/*      </SheetContent>*/}
+      {/*    </Sheet>*/}
+      {/*    <Breadcrumb className="hidden md:flex">*/}
+      {/*      <BreadcrumbList>*/}
+      {/*        <BreadcrumbItem>*/}
+      {/*          <BreadcrumbLink asChild>*/}
+      {/*            <Link href="#">Dashboard</Link>*/}
+      {/*          </BreadcrumbLink>*/}
+      {/*        </BreadcrumbItem>*/}
+      {/*        <BreadcrumbSeparator />*/}
+      {/*        <BreadcrumbItem>*/}
+      {/*          <BreadcrumbLink asChild>*/}
+      {/*            <Link href="#">Orders</Link>*/}
+      {/*          </BreadcrumbLink>*/}
+      {/*        </BreadcrumbItem>*/}
+      {/*        <BreadcrumbSeparator />*/}
+      {/*        <BreadcrumbItem>*/}
+      {/*          <BreadcrumbPage>Recent Orders</BreadcrumbPage>*/}
+      {/*        </BreadcrumbItem>*/}
+      {/*      </BreadcrumbList>*/}
+      {/*    </Breadcrumb>*/}
+      {/*    <div className="relative ml-auto flex-1 md:grow-0">*/}
+      {/*      <Search className="absolute left-2.5 top-2.5 h-4 w-4 text-muted-foreground" />*/}
+      {/*      <Input*/}
+      {/*        type="search"*/}
+      {/*        placeholder="Search..."*/}
+      {/*        className="w-full rounded-lg bg-background pl-8 md:w-[200px] lg:w-[336px]"*/}
+      {/*      />*/}
+      {/*    </div>*/}
+      {/*    <DropdownMenu>*/}
+      {/*      <DropdownMenuTrigger asChild>*/}
+      {/*        <Button*/}
+      {/*          variant="outline"*/}
+      {/*          size="icon"*/}
+      {/*          className="overflow-hidden rounded-full"*/}
+      {/*        >*/}
+      {/*          <Image*/}
+      {/*            src="/placeholder-user.jpg"*/}
+      {/*            width={36}*/}
+      {/*            height={36}*/}
+      {/*            alt="Avatar"*/}
+      {/*            className="overflow-hidden rounded-full"*/}
+      {/*          />*/}
+      {/*        </Button>*/}
+      {/*      </DropdownMenuTrigger>*/}
+      {/*      <DropdownMenuContent align="end">*/}
+      {/*        <DropdownMenuLabel>My Account</DropdownMenuLabel>*/}
+      {/*        <DropdownMenuSeparator />*/}
+      {/*        <DropdownMenuItem>Settings</DropdownMenuItem>*/}
+      {/*        <DropdownMenuItem>Support</DropdownMenuItem>*/}
+      {/*        <DropdownMenuSeparator />*/}
+      {/*        <DropdownMenuItem>Logout</DropdownMenuItem>*/}
+      {/*      </DropdownMenuContent>*/}
+      {/*    </DropdownMenu>*/}
+      {/*  </header>*/}
+      {/*  <main className="grid flex-1 items-start gap-4 p-4 sm:px-6 sm:py-0 md:gap-8 lg:grid-cols-3 xl:grid-cols-3">*/}
+      {/*    <div className="grid auto-rows-max items-start gap-4 md:gap-8 lg:col-span-2">*/}
+      {/*      <div className="grid gap-4 sm:grid-cols-2 md:grid-cols-4 lg:grid-cols-2 xl:grid-cols-4">*/}
+      {/*        <Card className="sm:col-span-2" x-chunk="dashboard-05-chunk-0">*/}
+      {/*          <CardHeader className="pb-3">*/}
+      {/*            <CardTitle>Your Orders</CardTitle>*/}
+      {/*            <CardDescription className="max-w-lg text-balance leading-relaxed">*/}
+      {/*              Introducing Our Dynamic Orders Dashboard for Seamless*/}
+      {/*              Management and Insightful Analysis.*/}
+      {/*            </CardDescription>*/}
+      {/*          </CardHeader>*/}
+      {/*          <CardFooter>*/}
+      {/*            <Button>Create New Order</Button>*/}
+      {/*          </CardFooter>*/}
+      {/*        </Card>*/}
+      {/*        <Card x-chunk="dashboard-05-chunk-1">*/}
+      {/*          <CardHeader className="pb-2">*/}
+      {/*            <CardDescription>This Week</CardDescription>*/}
+      {/*            <CardTitle className="text-4xl">$1,329</CardTitle>*/}
+      {/*          </CardHeader>*/}
+      {/*          <CardContent>*/}
+      {/*            <div className="text-xs text-muted-foreground">*/}
+      {/*              +25% from last week*/}
+      {/*            </div>*/}
+      {/*          </CardContent>*/}
+      {/*          <CardFooter>*/}
+      {/*            <Progress value={25} aria-label="25% increase" />*/}
+      {/*          </CardFooter>*/}
+      {/*        </Card>*/}
+      {/*        <Card x-chunk="dashboard-05-chunk-2">*/}
+      {/*          <CardHeader className="pb-2">*/}
+      {/*            <CardDescription>This Month</CardDescription>*/}
+      {/*            <CardTitle className="text-4xl">$5,329</CardTitle>*/}
+      {/*          </CardHeader>*/}
+      {/*          <CardContent>*/}
+      {/*            <div className="text-xs text-muted-foreground">*/}
+      {/*              +10% from last month*/}
+      {/*            </div>*/}
+      {/*          </CardContent>*/}
+      {/*          <CardFooter>*/}
+      {/*            <Progress value={12} aria-label="12% increase" />*/}
+      {/*          </CardFooter>*/}
+      {/*        </Card>*/}
+      {/*      </div>*/}
+      {/*      <Tabs defaultValue="week">*/}
+      {/*        <div className="flex items-center">*/}
+      {/*          <TabsList>*/}
+      {/*            <TabsTrigger value="week">Week</TabsTrigger>*/}
+      {/*            <TabsTrigger value="month">Month</TabsTrigger>*/}
+      {/*            <TabsTrigger value="year">Year</TabsTrigger>*/}
+      {/*          </TabsList>*/}
+      {/*          <div className="ml-auto flex items-center gap-2">*/}
+      {/*            <DropdownMenu>*/}
+      {/*              <DropdownMenuTrigger asChild>*/}
+      {/*                <Button*/}
+      {/*                  variant="outline"*/}
+      {/*                  size="sm"*/}
+      {/*                  className="h-7 gap-1 text-sm"*/}
+      {/*                >*/}
+      {/*                  <ListFilter className="h-3.5 w-3.5" />*/}
+      {/*                  <span className="sr-only sm:not-sr-only">Filter</span>*/}
+      {/*                </Button>*/}
+      {/*              </DropdownMenuTrigger>*/}
+      {/*              <DropdownMenuContent align="end">*/}
+      {/*                <DropdownMenuLabel>Filter by</DropdownMenuLabel>*/}
+      {/*                <DropdownMenuSeparator />*/}
+      {/*                <DropdownMenuCheckboxItem checked>*/}
+      {/*                  Fulfilled*/}
+      {/*                </DropdownMenuCheckboxItem>*/}
+      {/*                <DropdownMenuCheckboxItem>*/}
+      {/*                  Declined*/}
+      {/*                </DropdownMenuCheckboxItem>*/}
+      {/*                <DropdownMenuCheckboxItem>*/}
+      {/*                  Refunded*/}
+      {/*                </DropdownMenuCheckboxItem>*/}
+      {/*              </DropdownMenuContent>*/}
+      {/*            </DropdownMenu>*/}
+      {/*            <Button*/}
+      {/*              size="sm"*/}
+      {/*              variant="outline"*/}
+      {/*              className="h-7 gap-1 text-sm"*/}
+      {/*            >*/}
+      {/*              <File className="h-3.5 w-3.5" />*/}
+      {/*              <span className="sr-only sm:not-sr-only">Export</span>*/}
+      {/*            </Button>*/}
+      {/*          </div>*/}
+      {/*        </div>*/}
+      {/*        <TabsContent value="week">*/}
+      {/*          <Card x-chunk="dashboard-05-chunk-3">*/}
+      {/*            <CardHeader className="px-7">*/}
+      {/*              <CardTitle>Orders</CardTitle>*/}
+      {/*              <CardDescription>*/}
+      {/*                Recent orders from your store.*/}
+      {/*              </CardDescription>*/}
+      {/*            </CardHeader>*/}
+      {/*            <CardContent>*/}
+      {/*              <Table>*/}
+      {/*                <TableHeader>*/}
+      {/*                  <TableRow>*/}
+      {/*                    <TableHead>Customer</TableHead>*/}
+      {/*                    <TableHead className="hidden sm:table-cell">*/}
+      {/*                      Type*/}
+      {/*                    </TableHead>*/}
+      {/*                    <TableHead className="hidden sm:table-cell">*/}
+      {/*                      Status*/}
+      {/*                    </TableHead>*/}
+      {/*                    <TableHead className="hidden md:table-cell">*/}
+      {/*                      Date*/}
+      {/*                    </TableHead>*/}
+      {/*                    <TableHead className="text-right">Amount</TableHead>*/}
+      {/*                  </TableRow>*/}
+      {/*                </TableHeader>*/}
+      {/*                <TableBody>*/}
+      {/*                  <TableRow className="bg-accent">*/}
+      {/*                    <TableCell>*/}
+      {/*                      <div className="font-medium">Liam Johnson</div>*/}
+      {/*                      <div className="hidden text-sm text-muted-foreground md:inline">*/}
+      {/*                        liam@example.com*/}
+      {/*                      </div>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      Sale*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      <Badge className="text-xs" variant="secondary">*/}
+      {/*                        Fulfilled*/}
+      {/*                      </Badge>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden md:table-cell">*/}
+      {/*                      2023-06-23*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="text-right">$250.00</TableCell>*/}
+      {/*                  </TableRow>*/}
+      {/*                  <TableRow>*/}
+      {/*                    <TableCell>*/}
+      {/*                      <div className="font-medium">Olivia Smith</div>*/}
+      {/*                      <div className="hidden text-sm text-muted-foreground md:inline">*/}
+      {/*                        olivia@example.com*/}
+      {/*                      </div>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      Refund*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      <Badge className="text-xs" variant="outline">*/}
+      {/*                        Declined*/}
+      {/*                      </Badge>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden md:table-cell">*/}
+      {/*                      2023-06-24*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="text-right">$150.00</TableCell>*/}
+      {/*                  </TableRow>*/}
+      {/*                  <TableRow>*/}
+      {/*                    <TableCell>*/}
+      {/*                      <div className="font-medium">Noah Williams</div>*/}
+      {/*                      <div className="hidden text-sm text-muted-foreground md:inline">*/}
+      {/*                        noah@example.com*/}
+      {/*                      </div>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      Subscription*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      <Badge className="text-xs" variant="secondary">*/}
+      {/*                        Fulfilled*/}
+      {/*                      </Badge>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden md:table-cell">*/}
+      {/*                      2023-06-25*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="text-right">$350.00</TableCell>*/}
+      {/*                  </TableRow>*/}
+      {/*                  <TableRow>*/}
+      {/*                    <TableCell>*/}
+      {/*                      <div className="font-medium">Emma Brown</div>*/}
+      {/*                      <div className="hidden text-sm text-muted-foreground md:inline">*/}
+      {/*                        emma@example.com*/}
+      {/*                      </div>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      Sale*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      <Badge className="text-xs" variant="secondary">*/}
+      {/*                        Fulfilled*/}
+      {/*                      </Badge>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden md:table-cell">*/}
+      {/*                      2023-06-26*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="text-right">$450.00</TableCell>*/}
+      {/*                  </TableRow>*/}
+      {/*                  <TableRow>*/}
+      {/*                    <TableCell>*/}
+      {/*                      <div className="font-medium">Liam Johnson</div>*/}
+      {/*                      <div className="hidden text-sm text-muted-foreground md:inline">*/}
+      {/*                        liam@example.com*/}
+      {/*                      </div>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      Sale*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      <Badge className="text-xs" variant="secondary">*/}
+      {/*                        Fulfilled*/}
+      {/*                      </Badge>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden md:table-cell">*/}
+      {/*                      2023-06-23*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="text-right">$250.00</TableCell>*/}
+      {/*                  </TableRow>*/}
+      {/*                  <TableRow>*/}
+      {/*                    <TableCell>*/}
+      {/*                      <div className="font-medium">Liam Johnson</div>*/}
+      {/*                      <div className="hidden text-sm text-muted-foreground md:inline">*/}
+      {/*                        liam@example.com*/}
+      {/*                      </div>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      Sale*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      <Badge className="text-xs" variant="secondary">*/}
+      {/*                        Fulfilled*/}
+      {/*                      </Badge>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden md:table-cell">*/}
+      {/*                      2023-06-23*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="text-right">$250.00</TableCell>*/}
+      {/*                  </TableRow>*/}
+      {/*                  <TableRow>*/}
+      {/*                    <TableCell>*/}
+      {/*                      <div className="font-medium">Olivia Smith</div>*/}
+      {/*                      <div className="hidden text-sm text-muted-foreground md:inline">*/}
+      {/*                        olivia@example.com*/}
+      {/*                      </div>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      Refund*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      <Badge className="text-xs" variant="outline">*/}
+      {/*                        Declined*/}
+      {/*                      </Badge>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden md:table-cell">*/}
+      {/*                      2023-06-24*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="text-right">$150.00</TableCell>*/}
+      {/*                  </TableRow>*/}
+      {/*                  <TableRow>*/}
+      {/*                    <TableCell>*/}
+      {/*                      <div className="font-medium">Emma Brown</div>*/}
+      {/*                      <div className="hidden text-sm text-muted-foreground md:inline">*/}
+      {/*                        emma@example.com*/}
+      {/*                      </div>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      Sale*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden sm:table-cell">*/}
+      {/*                      <Badge className="text-xs" variant="secondary">*/}
+      {/*                        Fulfilled*/}
+      {/*                      </Badge>*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="hidden md:table-cell">*/}
+      {/*                      2023-06-26*/}
+      {/*                    </TableCell>*/}
+      {/*                    <TableCell className="text-right">$450.00</TableCell>*/}
+      {/*                  </TableRow>*/}
+      {/*                </TableBody>*/}
+      {/*              </Table>*/}
+      {/*            </CardContent>*/}
+      {/*          </Card>*/}
+      {/*        </TabsContent>*/}
+      {/*      </Tabs>*/}
+      {/*    </div>*/}
+      {/*    <div>*/}
+      {/*      <Card className="overflow-hidden" x-chunk="dashboard-05-chunk-4">*/}
+      {/*        <CardHeader className="flex flex-row items-start bg-muted/50">*/}
+      {/*          <div className="grid gap-0.5">*/}
+      {/*            <CardTitle className="group flex items-center gap-2 text-lg">*/}
+      {/*              Order Oe31b70H*/}
+      {/*              <Button*/}
+      {/*                size="icon"*/}
+      {/*                variant="outline"*/}
+      {/*                className="h-6 w-6 opacity-0 transition-opacity group-hover:opacity-100"*/}
+      {/*              >*/}
+      {/*                <Copy className="h-3 w-3" />*/}
+      {/*                <span className="sr-only">Copy Order ID</span>*/}
+      {/*              </Button>*/}
+      {/*            </CardTitle>*/}
+      {/*            <CardDescription>Date: November 23, 2023</CardDescription>*/}
+      {/*          </div>*/}
+      {/*          <div className="ml-auto flex items-center gap-1">*/}
+      {/*            <Button size="sm" variant="outline" className="h-8 gap-1">*/}
+      {/*              <Truck className="h-3.5 w-3.5" />*/}
+      {/*              <span className="lg:sr-only xl:not-sr-only xl:whitespace-nowrap">*/}
+      {/*                Track Order*/}
+      {/*              </span>*/}
+      {/*            </Button>*/}
+      {/*            <DropdownMenu>*/}
+      {/*              <DropdownMenuTrigger asChild>*/}
+      {/*                <Button size="icon" variant="outline" className="h-8 w-8">*/}
+      {/*                  <MoreVertical className="h-3.5 w-3.5" />*/}
+      {/*                  <span className="sr-only">More</span>*/}
+      {/*                </Button>*/}
+      {/*              </DropdownMenuTrigger>*/}
+      {/*              <DropdownMenuContent align="end">*/}
+      {/*                <DropdownMenuItem>Edit</DropdownMenuItem>*/}
+      {/*                <DropdownMenuItem>Export</DropdownMenuItem>*/}
+      {/*                <DropdownMenuSeparator />*/}
+      {/*                <DropdownMenuItem>Trash</DropdownMenuItem>*/}
+      {/*              </DropdownMenuContent>*/}
+      {/*            </DropdownMenu>*/}
+      {/*          </div>*/}
+      {/*        </CardHeader>*/}
+      {/*        <CardContent className="p-6 text-sm">*/}
+      {/*          <div className="grid gap-3">*/}
+      {/*            <div className="font-semibold">Order Details</div>*/}
+      {/*            <ul className="grid gap-3">*/}
+      {/*              <li className="flex items-center justify-between">*/}
+      {/*                <span className="text-muted-foreground">*/}
+      {/*                  Glimmer Lamps x <span>2</span>*/}
+      {/*                </span>*/}
+      {/*                <span>$250.00</span>*/}
+      {/*              </li>*/}
+      {/*              <li className="flex items-center justify-between">*/}
+      {/*                <span className="text-muted-foreground">*/}
+      {/*                  Aqua Filters x <span>1</span>*/}
+      {/*                </span>*/}
+      {/*                <span>$49.00</span>*/}
+      {/*              </li>*/}
+      {/*            </ul>*/}
+      {/*            <Separator className="my-2" />*/}
+      {/*            <ul className="grid gap-3">*/}
+      {/*              <li className="flex items-center justify-between">*/}
+      {/*                <span className="text-muted-foreground">Subtotal</span>*/}
+      {/*                <span>$299.00</span>*/}
+      {/*              </li>*/}
+      {/*              <li className="flex items-center justify-between">*/}
+      {/*                <span className="text-muted-foreground">Shipping</span>*/}
+      {/*                <span>$5.00</span>*/}
+      {/*              </li>*/}
+      {/*              <li className="flex items-center justify-between">*/}
+      {/*                <span className="text-muted-foreground">Tax</span>*/}
+      {/*                <span>$25.00</span>*/}
+      {/*              </li>*/}
+      {/*              <li className="flex items-center justify-between font-semibold">*/}
+      {/*                <span className="text-muted-foreground">Total</span>*/}
+      {/*                <span>$329.00</span>*/}
+      {/*              </li>*/}
+      {/*            </ul>*/}
+      {/*          </div>*/}
+      {/*          <Separator className="my-4" />*/}
+      {/*          <div className="grid grid-cols-2 gap-4">*/}
+      {/*            <div className="grid gap-3">*/}
+      {/*              <div className="font-semibold">Shipping Information</div>*/}
+      {/*              <address className="grid gap-0.5 not-italic text-muted-foreground">*/}
+      {/*                <span>Liam Johnson</span>*/}
+      {/*                <span>1234 Main St.</span>*/}
+      {/*                <span>Anytown, CA 12345</span>*/}
+      {/*              </address>*/}
+      {/*            </div>*/}
+      {/*            <div className="grid auto-rows-max gap-3">*/}
+      {/*              <div className="font-semibold">Billing Information</div>*/}
+      {/*              <div className="text-muted-foreground">*/}
+      {/*                Same as shipping address*/}
+      {/*              </div>*/}
+      {/*            </div>*/}
+      {/*          </div>*/}
+      {/*          <Separator className="my-4" />*/}
+      {/*          <div className="grid gap-3">*/}
+      {/*            <div className="font-semibold">Customer Information</div>*/}
+      {/*            <dl className="grid gap-3">*/}
+      {/*              <div className="flex items-center justify-between">*/}
+      {/*                <dt className="text-muted-foreground">Customer</dt>*/}
+      {/*                <dd>Liam Johnson</dd>*/}
+      {/*              </div>*/}
+      {/*              <div className="flex items-center justify-between">*/}
+      {/*                <dt className="text-muted-foreground">Email</dt>*/}
+      {/*                <dd>*/}
+      {/*                  <a href="mailto:">liam@acme.com</a>*/}
+      {/*                </dd>*/}
+      {/*              </div>*/}
+      {/*              <div className="flex items-center justify-between">*/}
+      {/*                <dt className="text-muted-foreground">Phone</dt>*/}
+      {/*                <dd>*/}
+      {/*                  <a href="tel:">+1 234 567 890</a>*/}
+      {/*                </dd>*/}
+      {/*              </div>*/}
+      {/*            </dl>*/}
+      {/*          </div>*/}
+      {/*          <Separator className="my-4" />*/}
+      {/*          <div className="grid gap-3">*/}
+      {/*            <div className="font-semibold">Payment Information</div>*/}
+      {/*            <dl className="grid gap-3">*/}
+      {/*              <div className="flex items-center justify-between">*/}
+      {/*                <dt className="flex items-center gap-1 text-muted-foreground">*/}
+      {/*                  <CreditCard className="h-4 w-4" />*/}
+      {/*                  Visa*/}
+      {/*                </dt>*/}
+      {/*                <dd>**** **** **** 4532</dd>*/}
+      {/*              </div>*/}
+      {/*            </dl>*/}
+      {/*          </div>*/}
+      {/*        </CardContent>*/}
+      {/*        <CardFooter className="flex flex-row items-center border-t bg-muted/50 px-6 py-3">*/}
+      {/*          <div className="text-xs text-muted-foreground">*/}
+      {/*            Updated <time dateTime="2023-11-23">November 23, 2023</time>*/}
+      {/*          </div>*/}
+      {/*          <Pagination className="ml-auto mr-0 w-auto">*/}
+      {/*            <PaginationContent>*/}
+      {/*              <PaginationItem>*/}
+      {/*                <Button size="icon" variant="outline" className="h-6 w-6">*/}
+      {/*                  <ChevronLeft className="h-3.5 w-3.5" />*/}
+      {/*                  <span className="sr-only">Previous Order</span>*/}
+      {/*                </Button>*/}
+      {/*              </PaginationItem>*/}
+      {/*              <PaginationItem>*/}
+      {/*                <Button size="icon" variant="outline" className="h-6 w-6">*/}
+      {/*                  <ChevronRight className="h-3.5 w-3.5" />*/}
+      {/*                  <span className="sr-only">Next Order</span>*/}
+      {/*                </Button>*/}
+      {/*              </PaginationItem>*/}
+      {/*            </PaginationContent>*/}
+      {/*          </Pagination>*/}
+      {/*        </CardFooter>*/}
+      {/*      </Card>*/}
+      {/*    </div>*/}
+      {/*  </main>*/}
+      {/*</div>*/}
+    </div>
   )
 }
